Added Config for the Energy Cell System
