+++ src/main/java/net/origamiking/mcmods/orm/client/EnergyCellsHudOverlay.java
@@ -18,34 +18,36 @@ public class EnergyCellsHudOverlay implements HudRenderCallback {
 
     @Override
     public void onHudRender(DrawContext drawContext, float tickDelta) {
-        ClientPlayerEntity player = MinecraftClient.getInstance().player;
-        if (ArmorUtils.isArmorSetOfType(player, TransformerArmorItem.class) && !player.isSpectator() && !player.getAbilities().creativeMode) {
-            int x = 0;
-            int y = 0;
-            MinecraftClient client = MinecraftClient.getInstance();
-            if (client != null) {
-                int width = client.getWindow().getScaledWidth();
-                int height = client.getWindow().getScaledHeight();
+        if (!OrmMain.getOrmConfig().disableEnergyCells) {
+            ClientPlayerEntity player = MinecraftClient.getInstance().player;
+            if (ArmorUtils.isArmorSetOfType(player, TransformerArmorItem.class) && !player.isSpectator() && !player.getAbilities().creativeMode) {
+                int x = 0;
+                int y = 0;
+                MinecraftClient client = MinecraftClient.getInstance();
+                if (client != null) {
+                    int width = client.getWindow().getScaledWidth();
+                    int height = client.getWindow().getScaledHeight();
 
-                x = width / 2;
-                y = height;
-            }
-            int texture_size = 12;
-            int number_of_cells = 10;
+                    x = width / 2;
+                    y = height;
+                }
+                int texture_size = 12;
+                int number_of_cells = 10;
 
-            RenderSystem.setShader(GameRenderer::getPositionTexProgram);
-            RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-            RenderSystem.setShaderTexture(0, EMPTY_ENERGY_CELL);
-            for (int e = 0; e < number_of_cells; e++) {
-                drawContext.drawTexture(EMPTY_ENERGY_CELL, x + 95 + (e * number_of_cells - 1), y - 18, 0, 0, texture_size, texture_size, texture_size, texture_size);
-            }
+                RenderSystem.setShader(GameRenderer::getPositionTexProgram);
+                RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+                RenderSystem.setShaderTexture(0, EMPTY_ENERGY_CELL);
+                for (int e = 0; e < number_of_cells; e++) {
+                    drawContext.drawTexture(EMPTY_ENERGY_CELL, x + 95 + (e * number_of_cells - 1), y - 18, 0, 0, texture_size, texture_size, texture_size, texture_size);
+                }
 
-            RenderSystem.setShaderTexture(0, FILLED_ENERGY_CELL);
-            for (int f = 0; f < number_of_cells; f++) {
-                if (((IEntityEnergyCellsDataSaver) player).getPersistentData().getInt("energy_cells") > f) {
-                    drawContext.drawTexture(FILLED_ENERGY_CELL, x + 95 + (f * number_of_cells - 1), y - 18, 0, 0, texture_size, texture_size, texture_size, texture_size);
-                } else {
-                    break;
+                RenderSystem.setShaderTexture(0, FILLED_ENERGY_CELL);
+                for (int f = 0; f < number_of_cells; f++) {
+                    if (((IEntityEnergyCellsDataSaver) player).getPersistentData().getInt("energy_cells") > f) {
+                        drawContext.drawTexture(FILLED_ENERGY_CELL, x + 95 + (f * number_of_cells - 1), y - 18, 0, 0, texture_size, texture_size, texture_size, texture_size);
+                    } else {
+                        break;
+                    }
                 }
             }
         }
