+++ src/main/java/net/origamiking/mcmods/orm/commands/energy_cell_commands/EnergyCellCommands.java
@@ -10,6 +10,8 @@ import net.minecraft.command.argument.EntityArgumentType;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.text.Text;
+import net.origamiking.mcmods.orm.OrmMain;
 import net.origamiking.mcmods.orm.utils.EnergyCellsData;
 import net.origamiking.mcmods.orm.utils.IEntityEnergyCellsDataSaver;
 
@@ -18,9 +20,10 @@ import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
 public class EnergyCellCommands {
     public static void register() {
         CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) ->
-            make(dispatcher)
+                make(dispatcher)
         );
     }
+
     private static void make(CommandDispatcher<ServerCommandSource> dispatcher) {
         LiteralArgumentBuilder<ServerCommandSource> command = LiteralArgumentBuilder.literal("orm-energy-cells");
 
@@ -44,20 +47,28 @@ public class EnergyCellCommands {
     }
 
     private static int executeAddEnergyCells(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
-        ServerPlayerEntity player = EntityArgumentType.getPlayer(context, "player");
-        int amount = IntegerArgumentType.getInteger(context, "amount");
+        if (!OrmMain.getOrmConfig().disableEnergyCells) {
+            ServerPlayerEntity player = EntityArgumentType.getPlayer(context, "player");
+            int amount = IntegerArgumentType.getInteger(context, "amount");
 
-        EnergyCellsData.addEnergyCells((IEntityEnergyCellsDataSaver) player, amount);
-
+            EnergyCellsData.addEnergyCells((IEntityEnergyCellsDataSaver) player, amount);
+            (context.getSource()).sendMessage(Text.translatable("message.orm.added_energy_cell"));
+        } else {
+            (context.getSource()).sendMessage(Text.translatable("message.orm.disabled_energy_cell"));
+        }
         return SINGLE_SUCCESS;
     }
 
     private static int executeRemoveEnergyCells(CommandContext<ServerCommandSource> context) throws CommandSyntaxException {
-        ServerPlayerEntity player = EntityArgumentType.getPlayer(context, "player");
-        int amount = IntegerArgumentType.getInteger(context, "amount");
+        if (!OrmMain.getOrmConfig().disableEnergyCells) {
+            ServerPlayerEntity player = EntityArgumentType.getPlayer(context, "player");
+            int amount = IntegerArgumentType.getInteger(context, "amount");
 
-        EnergyCellsData.removeEnergyCells((IEntityEnergyCellsDataSaver) player, amount);
-
+            EnergyCellsData.removeEnergyCells((IEntityEnergyCellsDataSaver) player, amount);
+            (context.getSource()).sendMessage(Text.translatable("message.orm.removed_energy_cell"));
+        } else {
+            (context.getSource()).sendMessage(Text.translatable("message.orm.disabled_energy_cell"));
+        }
         return SINGLE_SUCCESS;
     }
 }
