+++ src/main/java/net/origamiking/mcmods/orm/config/ConfigDefaultValues.java
@@ -17,4 +17,7 @@ public class ConfigDefaultValues {
     protected static int defaultTransformerBootsProtectionValue = 3;
     protected static int defaultTransformerArmorEnchantability = 20;
     protected static boolean defaultEnableTransforming = true;
+    protected static boolean defaultEnergyCells = false;
+    public static float defaultChanceToRemoveEnergyCell = 0.000005f;
+
 }
