+++ src/main/java/net/origamiking/mcmods/orm/items/custom/energy_cell/EnergyCellItem.java
@@ -9,6 +9,7 @@ import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.world.World;
 import net.origamiking.mcmods.oapi.armor.ArmorUtils;
+import net.origamiking.mcmods.orm.OrmMain;
 import net.origamiking.mcmods.orm.networking.ModMessages;
 import net.origamiking.mcmods.orm.utils.IEntityEnergyCellsDataSaver;
 import net.origamiking.mcmods.orm.utils.TransformerArmorItem;
@@ -22,14 +23,18 @@ public class EnergyCellItem extends Item {
     @Override
     public TypedActionResult<ItemStack> use(World world, PlayerEntity player, Hand hand) {
         //on both Client and Server
-        if (ArmorUtils.isArmorSetOfType(player, TransformerArmorItem.class) && !(((IEntityEnergyCellsDataSaver) player).getPersistentData().getInt("energy_cells") >= 10)) {
-            if (world.isClient()) {
-                ClientPlayNetworking.send(ModMessages.ENERGY_CELLS_ID, PacketByteBufs.create());
+        if (!OrmMain.getOrmConfig().disableEnergyCells) {
+            if (ArmorUtils.isArmorSetOfType(player, TransformerArmorItem.class) && !(((IEntityEnergyCellsDataSaver) player).getPersistentData().getInt("energy_cells") >= 10)) {
+                if (world.isClient()) {
+                    ClientPlayNetworking.send(ModMessages.ENERGY_CELLS_ID, PacketByteBufs.create());
+                }
+                if (!player.getAbilities().creativeMode) {
+                    player.getStackInHand(hand).decrement(1);
+                }
+                return TypedActionResult.consume(player.getStackInHand(hand));
+            } else {
+                return TypedActionResult.pass(player.getStackInHand(hand));
             }
-            if (!player.getAbilities().creativeMode) {
-                player.getStackInHand(hand).decrement(1);
-            }
-            return TypedActionResult.consume(player.getStackInHand(hand));
         } else {
             return TypedActionResult.pass(player.getStackInHand(hand));
         }
